requirements.txt
Flask==3.0.3
requests==2.32.3
polyline==2.0.2
shapely==2.0.5
geopandas==1.0.1
python-dotenv==1.0.1
